moved specs to relevent package
